Set up immobilienkauf card layout
edit html to display Immobilienkauf-cards in such a way,
that it doesn't interfere with the classes of other html files.
Namespace shouldn't be poluted in any way, so that JS can modify
pages on the fly.

Structure the css file in a way, that different styles for different
pseudo pages are easy to find.